src/test/java/ru/yandex/practicum/filmorate/storageTest/FilmDbStorageTest.java
package ru.yandex.practicum.filmorate.storageTest;

import lombok.RequiredArgsConstructor;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.context.SpringBootTest;
import ru.yandex.practicum.filmorate.model.Film;
import ru.yandex.practicum.filmorate.model.Mpa;
import ru.yandex.practicum.filmorate.storage.film.FilmDbStorage;
import java.time.LocalDate;
import java.util.Optional;
import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
@AutoConfigureTestDatabase
@RequiredArgsConstructor(onConstructor_ = @Autowired)
public class FilmDbStorageTest {
    private final FilmDbStorage filmStorage;

    @Test
    void shouldReadById() {
        Film testFilm = Film.builder()
                .name("name")
                .description("description")
                .duration(150)
                .releaseDate(LocalDate.parse("2000-01-01"))
                .mpa(Mpa.builder().id(1).build())
                .build();

        Long filmId = filmStorage.create(testFilm).getId();

        Optional<Film> filmOptional = Optional.ofNullable(filmStorage.readById(filmId));
        assertThat(filmOptional)
                .isPresent()
                .hasValueSatisfying(film ->
                        assertThat(film).hasFieldOrPropertyWithValue("id", filmId)
                );
    }


}
